BF-69: added Gyms Near Me
