--- lib/blocs/article/article_event.dart
@@ -1,10 +1,0 @@
-part of 'article_bloc.dart';
-
-abstract class ArticleEvent extends Equatable {
-  const ArticleEvent();
-
-  @override
-  List<Object> get props => [];
-}
-
-class LoadArticle extends ArticleEvent {}
