--- lib/blocs/gym/gym_event.dart
@@ -1,10 +1,0 @@
-part of 'gym_bloc.dart';
-
-abstract class GymEvent extends Equatable {
-  const GymEvent();
-
-  @override
-  List<Object> get props => [];
-}
-
-class LoadGym extends GymEvent {}
