--- lib/views/home/home_area/welcome_card.dart
@@ -1,1 +1,0 @@
-
