+++ lib/views/main/home_screen/home_screen.dart
@@ -1,5 +1,6 @@
 import 'package:bfit_tracker/controllers/home_controller.dart';
 import 'package:bfit_tracker/views/main/home_screen/widgets/article_list.dart';
+import 'package:bfit_tracker/views/main/home_screen/widgets/gyms_near_me_list.dart';
 import 'package:bfit_tracker/views/main/home_screen/widgets/welcome_card.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
@@ -9,19 +10,23 @@ class HomeScreen extends GetView<HomeController> {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        WelcomeCard(),
-        Padding(
-          padding: const EdgeInsets.symmetric(vertical: 8.0),
-          child: Obx(
-            () => controller.articles == null || controller.articles.isEmpty
-                ? CircularProgressIndicator()
-                : ArticleList()
+    return SingleChildScrollView(
+      child: Column(
+        children: [
+          WelcomeCard(),
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8),
+            child: Obx(() =>
+                controller.articles == null || controller.articles.isEmpty
+                    ? CircularProgressIndicator()
+                    : ArticleList()),
           ),
-        ),
-      ],
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8),
+            child: GymsNearMeList(),
+          ),
+        ],
+      ),
     );
   }
 }
