+++ pubspec.yaml
@@ -34,7 +34,7 @@ dependencies:
   flutter_bloc: ^4.0.0
   flutter_dotenv: ^2.1.0
   font_awesome_flutter: ^8.10.0
-  geolocator: ^7.0.1
+  geolocator: ^7.0.3
   get: ^4.0.0-nullsafety.2
   googleapis: ^2.0.0
   google_sign_in: ^4.5.3
